refactor(core/http): decouple Liferay-specific types and config from core/http
- Rename LiferayApiClient -> HttpApiClient in core/http/client.ts
- Rename LiferayAuthConfig -> OAuthClientCredentialsConfig in core/http/auth.ts
- Move resolveLiferayConfig from core/http/config.ts to
  src/features/liferay/liferay-connection-config.ts (fixes inverted dep)
- Replace LIFERAY_HTTP_ERROR -> HTTP_ERROR in core/http/client.ts
- Replace LIFERAY_AUTH_ERROR -> AUTH_ERROR in core/http/auth.ts
- Replace LIFERAY_AUTH_CONFIG_ERROR -> AUTH_CONFIG_ERROR in core/http/auth.ts
- Fix redundant import path ../../core/errors.js -> ../errors.js in auth.ts
- Update config-builder.ts to import resolveLiferayConfig from features/liferay
- Keep core/http/config.ts as re-export shim for backward compat
