+++ src/features/liferay/content/liferay-content-journal-shared.ts
@@ -1,7 +1,7 @@
 import type {AppConfig} from '../../../core/config/load-config.js';
 import {CliError} from '../../../core/errors.js';
 import type {OAuthTokenClient} from '../../../core/http/auth.js';
-import type {LiferayApiClient} from '../../../core/http/client.js';
+import type {HttpApiClient} from '../../../core/http/client.js';
 import type {LiferayGateway} from '../liferay-gateway.js';
 import {LiferayErrors} from '../errors/index.js';
 import {fetchPagedItems, normalizeLocalizedName} from '../inventory/liferay-inventory-shared.js';
@@ -31,7 +31,7 @@ export type JsonwsJournalFolder = {
 
 export async function resolveJournalStructureDefinitions(
   config: AppConfig,
-  apiClient: LiferayApiClient,
+  apiClient: HttpApiClient,
   tokenClient: OAuthTokenClient,
   groupId: number,
   structureMap: Map<number, string>,
