+++ src/features/liferay/liferay-audit.ts
@@ -1,6 +1,6 @@
 import type {AppConfig} from '../../core/config/load-config.js';
 import {createOAuthTokenClient, type OAuthTokenClient} from '../../core/http/auth.js';
-import {createLiferayApiClient, type LiferayApiClient} from '../../core/http/client.js';
+import {createLiferayApiClient, type HttpApiClient} from '../../core/http/client.js';
 import {fetchPagedItems, resolveSite} from './inventory/liferay-inventory-shared.js';
 import {createLiferayGateway} from './liferay-gateway.js';
 import {performLiferayHealthCheck} from './liferay-health.js';
@@ -22,7 +22,7 @@ export type LiferayAuditResult = {
 };
 
 type AuditDependencies = {
-  apiClient?: LiferayApiClient;
+  apiClient?: HttpApiClient;
   tokenClient?: OAuthTokenClient;
 };
