+++ src/features/liferay/liferay-preflight.ts
@@ -12,7 +12,7 @@
 import type {AppConfig} from '../../core/config/load-config.js';
 import type {OAuthTokenClient} from '../../core/http/auth.js';
 import {createOAuthTokenClient} from '../../core/http/auth.js';
-import type {LiferayApiClient} from '../../core/http/client.js';
+import type {HttpApiClient} from '../../core/http/client.js';
 import {createLiferayApiClient} from '../../core/http/client.js';
 import {buildAuthOptions} from './liferay-http-shared.js';
 import {LookupCache} from './lookup-cache.js';
@@ -35,7 +35,7 @@ export type PreflightResult = {
 };
 
 type PreflightDependencies = {
-  apiClient?: LiferayApiClient;
+  apiClient?: HttpApiClient;
   tokenClient?: OAuthTokenClient;
   accessToken?: string;
   forceRefresh?: boolean;
@@ -46,7 +46,7 @@ const preflightCache = new LookupCache<PreflightResult>();
 
 async function probeUrl(
   config: AppConfig,
-  apiClient: LiferayApiClient,
+  apiClient: HttpApiClient,
   accessToken: string,
   path: string,
 ): Promise<SurfaceStatus> {
