+++ src/features/liferay/page-layout/liferay-page-layout-diff.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 
 import type {AppConfig} from '../../../core/config/load-config.js';
 import type {OAuthTokenClient} from '../../../core/http/auth.js';
-import type {LiferayApiClient} from '../../../core/http/client.js';
+import type {HttpApiClient} from '../../../core/http/client.js';
 import {LiferayErrors} from '../errors/index.js';
 import {runLiferayPageLayoutExport, type LiferayPageLayoutExport} from './liferay-page-layout-export.js';
 
@@ -12,7 +12,7 @@ const EXPORT_KIND = 'liferay-page-layout-export';
 const MAX_DIFFS = 50;
 
 type PageLayoutDiffDependencies = {
-  apiClient?: LiferayApiClient;
+  apiClient?: HttpApiClient;
   tokenClient?: OAuthTokenClient;
   now?: () => Date;
 };
