+++ src/features/liferay/resource/liferay-resource-get-structure.ts
@@ -1,6 +1,6 @@
 import type {AppConfig} from '../../../core/config/load-config.js';
 import type {OAuthTokenClient} from '../../../core/http/auth.js';
-import type {LiferayApiClient} from '../../../core/http/client.js';
+import type {HttpApiClient} from '../../../core/http/client.js';
 import {LiferayErrors} from '../errors/index.js';
 import {createLiferayGateway} from '../liferay-gateway.js';
 import {normalizeLocalizedName} from '../inventory/liferay-inventory-shared.js';
@@ -8,7 +8,7 @@ import {buildResourceSiteChain, resolveResourceSite} from './liferay-resource-sh
 import type {DataDefinitionPayload} from './liferay-resource-payloads.js';
 
 type ResourceDependencies = {
-  apiClient?: LiferayApiClient;
+  apiClient?: HttpApiClient;
   tokenClient?: OAuthTokenClient;
 };
