+++ tests/unit/liferay-resource-sync-shared.test.ts
@@ -2,7 +2,7 @@ import {describe, expect, test, vi} from 'vitest';
 
 import type {AppConfig} from '../../src/core/config/load-config.js';
 import type {OAuthTokenClient, TokenResponse} from '../../src/core/http/auth.js';
-import type {HttpResponse, LiferayApiClient} from '../../src/core/http/client.js';
+import type {HttpResponse, HttpApiClient} from '../../src/core/http/client.js';
 import {postFormCandidates} from '../../src/features/liferay/resource/liferay-resource-sync-shared.js';
 
 const mockConfig: AppConfig = {
@@ -29,7 +29,7 @@ const mockToken: TokenResponse = {
   expiresIn: 3600,
 };
 
-const createMockApiClient = (): LiferayApiClient => ({
+const createMockApiClient = (): HttpApiClient => ({
   get: vi.fn(),
   delete: vi.fn(),
   postJson: vi.fn(),
